More adjustments and changes needed for C++20

+++ src/engraving/rendering/score/tdraw.cpp
@@ -1272,7 +1272,9 @@ void TDraw::draw(const FiguredBassItem* item, Painter* painter, const PaintOptio
     painter->setBrush(BrushStyle::NoBrush);
     Pen pen(item->figuredBass()->curColor(opt), FiguredBass::FB_CONTLINE_THICKNESS * _spatium, PenStyle::SolidLine, PenCapStyle::RoundCap);
     painter->setPen(pen);
-    painter->drawText(ldata->bbox(), muse::draw::TextDontClip | muse::draw::AlignLeft | muse::draw::AlignTop, ldata->displayText);
+    painter->drawText(ldata->bbox(),
+                      static_cast<int>(muse::draw::TextDontClip) | static_cast<int>(muse::draw::AlignLeft | muse::draw::AlignTop),
+                      ldata->displayText);
 
     // continuation line
     double lineEndX = 0.0;
@@ -1489,7 +1491,7 @@ void TDraw::draw(const FretDiagram* item, Painter* painter, const PaintOptions& 
             painter->rotate(90);
             if (item->numPos() == 0) {
                 painter->drawText(RectF(.0, ldata->stringDist * (item->strings() - 1), .0, .0),
-                                  muse::draw::AlignLeft | muse::draw::TextDontClip, text);
+                                  static_cast<int>(muse::draw::AlignLeft) | static_cast<int>(muse::draw::TextDontClip), text);
             } else {
                 painter->drawText(RectF(.0, .0, .0, .0),
                                   muse::draw::AlignBottom | muse::draw::AlignLeft | muse::draw::TextDontClip, text);
